box: split out components
