+++ src/serlo-editor/plugins/box/toolbar.tsx
@@ -25,8 +25,8 @@ export const BoxToolbar = ({ id, state }: BoxProps) => {
             <select
               onChange={(e) => state.type.set(e.target.value)}
               className={tw`
-                mr-2 cursor-pointer rounded-md !border border-gray-500 bg-editor-primary-100 px-1 py-[1px] text-sm transition-all
-                hover:bg-editor-primary-200 focus:bg-editor-primary-200 focus:outline-none
+                mr-2 cursor-pointer rounded-md !border border-gray-500 bg-editor-primary-100 px-1 py-[1px]
+                text-sm transition-all hover:bg-editor-primary-200 focus:bg-editor-primary-200 focus:outline-none
               `}
               value={state.type.value}
             >
